separated react and redux into folders
