+++ src/App.test.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import ReactDOM from "react-dom";
 import App from "./App";
 import { Provider } from "react-redux";
-import createReduxStore from "./store/store";
-import initialState from "./store/initialState";
-import ApplicationState from "./store/applicationState";
+import createReduxStore from "./redux/store/store";
+import initialState from "./redux/store/initialState";
+import ApplicationState from "./redux/store/applicationState";
 
 it("renders without crashing", () => {
     const defaultState: ApplicationState = initialState;
