+++ src/react/components/common/condensedList.scss
@@ -1,4 +1,4 @@
-@import '../../assets/sass/theme.scss';
+@import '../../../assets/sass/theme.scss';
 
 .condensed-list {
