+++ src/react/components/common/condensedList.tsx
