+++ src/react/components/common/connectionPicker.test.tsx
