+++ src/react/components/common/tagsInput.scss
