+++ src/react/components/common/tagsInput.tsx
