+++ src/react/components/pages/appSettings/appSettings.json
