+++ src/react/components/pages/appSettings/appSettings.scss
