+++ src/react/components/pages/appSettings/appSettings.ui.json
