+++ src/react/components/pages/appSettings/appSettingsPage.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import { connect } from "react-redux";
 import { bindActionCreators } from "redux";
 import deepmerge from "deepmerge";
-import IApplicationActions, * as applicationActions from "../../../actions/applicationActions";
-import IConnectionActions, * as connectionActions from "../../../actions/connectionActions";
-import ApplicationState, { IAppSettings, IConnection } from "../../../store/applicationState";
+import IApplicationActions, * as applicationActions from "../../../../redux/actions/applicationActions";
+import IConnectionActions, * as connectionActions from "../../../../redux/actions/connectionActions";
+import ApplicationState, { IAppSettings, IConnection } from "../../../../redux/store/applicationState";
 import Form from "react-jsonschema-form";
 import formSchema from "./appSettings.json";
 import uiSchema from "./appSettings.ui.json";
