+++ src/react/components/pages/connections/connectionItem.tsx
