+++ src/react/components/pages/connections/connectionsPage.scss
