+++ src/react/components/pages/connections/connectionsPage.tsx
@@ -4,9 +4,9 @@ import { connect } from "react-redux";
 import { bindActionCreators } from "redux";
 import ConnectionItem from "./connectionItem";
 import CondensedList from "../../common/condensedList";
-import ApplicationState, { IConnection } from "../../../store/applicationState.js";
+import ApplicationState, { IConnection } from "../../../../redux/store/applicationState.js";
 import { RouteComponentProps } from "react-router-dom";
-import IConnectionActions, * as connectionActions from "../../../actions/connectionActions";
+import IConnectionActions, * as connectionActions from "../../../../redux/actions/connectionActions";
 import ConnectionForm from "./connectionForm";
 import "./connectionsPage.scss";
 import { AzureCloudStorageService } from "../../../providers/storage/azureBlobStorage";
