+++ src/react/components/pages/homepage/homePage.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/sass/theme.scss';
+@import '../../../../assets/sass/theme.scss';
 
 .app-homepage {
     height: 100%;
