+++ src/react/components/pages/homepage/homePage.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import { connect } from "react-redux";
 import { bindActionCreators } from "redux";
 import "./homePage.scss";
-import IProjectActions, * as projectActions from "../../../actions/projectActions";
-import ApplicationState, { IProject } from "../../../store/applicationState";
+import IProjectActions, * as projectActions from "../../../../redux/actions/projectActions";
+import ApplicationState, { IProject } from "../../../../redux/store/applicationState";
 import CondensedList from "../../common/condensedList";
 import RecentProjectItem from "./recentProjectItem";
 import FilePicker from "../../common/filePicker";
