+++ src/react/components/shell/sidebar.test.tsx
