+++ src/react/components/shell/sidebar.tsx
