+++ src/redux/reducers/applicationReducer.ts
