+++ src/redux/reducers/connectionsReducer.ts
