+++ src/redux/reducers/currentProjectReducer.ts
